+++ dev/examples/inc/dynamic-features.ts
@@ -1,0 +1,3 @@
+import { AnimationFeature } from "@framer"
+
+export default [AnimationFeature]